refactor(tests): name fixture suites by source
