+++ AGENTS.md
@@ -20,7 +20,7 @@ This is a C++ desktop app built with CMake and relying on static libraries like 
 
 - `QtProject/app`: main application code. `MainWindow` handles scanning/progress, `Video` extracts metadata/thumbnails/hashes, `Comparison` reviews matches and cleanup actions, `Db` owns cache persistence.
 - `QtProject/app/*.ui`: Qt Designer forms with auto-connected `on_<object>_<signal>` slots. Keep UI changes consistent with the generated `ui_*.h` flow.
-- `QtProject/tests`: `test_repo_*` targets are self-contained and use only checked-in fixtures. Optional corpora live under `QtProject/tests/local` and use flat `test_local_*` target names. `test_repo_auto_delete` covers focused end-to-end cleanup; `test_repo_video_extraction_regression` snapshots the Nice videos; and `test_repo_video_matching` validates the tracked matching manifest.
+- `QtProject/tests`: `test_repo_*` targets are self-contained and use only checked-in fixtures under `QtProject/tests/repo`. Optional corpora live under `QtProject/tests/external` and use flat `test_external_*` target names. `test_repo_auto_delete` covers focused end-to-end cleanup; `test_repo_video_extraction_regression` snapshots the Nice videos; and `test_repo_video_matching` validates the tracked matching manifest.
 - `samples/videos`: small representative fixtures for video-processing tests. Avoid replacing binary fixtures unless needed for the test intent.
 - Keep repository-tracked video fixtures within a few MB total because every clone downloads them.
 - Keep the out of repo, optional `~/Dev` video corpus lean as well; a few GB total is acceptable.
@@ -36,13 +36,13 @@ The main development platform is macOS; keep default agent commands on this path
   `cmake --build QtProject/builds/build-debug-6.10.1-macos --target test_repo_auto_delete && ctest --test-dir QtProject/builds/build-debug-6.10.1-macos -C Debug --output-on-failure -R ^test_repo_auto_delete$`
 - Run the self-contained CTest baseline:
   `ctest --test-dir QtProject/builds/build-debug-6.10.1-macos -C Debug --output-on-failure -R "^(test_comparison|test_mainwindow|test_repo_auto_delete|test_repo_video_matching)$"`
-- CTest labels make the safe lanes explicit: `ctest --test-dir QtProject/builds/build-debug-6.10.1-macos -L repo-fixtures` runs tracked fixtures, while `-L local-fixtures` runs the optional `~/Dev` suites. `test_local_large_video_corpus` is separately labeled `external-corpus` and `requires-mounted-100gb`; run only its named functions.
-- Green `test_local_whole_app_scan` function cases for regular local-corpus work: `emptyDb`, `test_whole_app_nocache`, `test_whole_app_cached`, `test_whole_app_cache_only`.
+- CTest labels make the safe lanes explicit: `ctest --test-dir QtProject/builds/build-debug-6.10.1-macos -L repo-fixtures` runs tracked fixtures, while `-L external-fixtures` runs the optional `~/Dev` suites. `test_external_large_video_corpus` is separately labeled `external-corpus` and `requires-mounted-100gb`; run only its named functions.
+- Green `test_external_whole_app_scan` function cases for regular external-corpus work: `emptyDb`, `test_whole_app_nocache`, `test_whole_app_cached`, `test_whole_app_cache_only`.
 - `test_repo_video_extraction_regression` compares platform-sensitive metadata/thumbnails for the Nice videos; run it on the macOS dev setup, not Linux CI, unless refreshing reference expectations.
-- `test_local_video_extraction_regression` compares every `~/Dev` video against its metadata and thumbnail references in each cache mode. Reference-detail cases (`test_check_refvidparams_nocache`, `test_check_refvidparams_withcache`, `test_check_refvidparams_withCacheOnly`) are not all green currently; run them when touching metadata, thumbnails, cache behavior, or reference data.
-- `test_local_whole_app_scan` runs `~/Dev` end-to-end scans in each cache mode; run its explicit functions only when that corpus is available.
-- To investigate one `~/Dev` extraction reference, run a single `test_local_video_extraction_regression` data row, for example `test_check_refvidparams_nocache:20150727_115225.mp4`.
+- `test_external_video_extraction_regression` compares every `~/Dev` video against its metadata and thumbnail references in each cache mode. Reference-detail cases (`test_check_refvidparams_nocache`, `test_check_refvidparams_withcache`, `test_check_refvidparams_withCacheOnly`) are not all green currently; run them when touching metadata, thumbnails, cache behavior, or reference data.
+- `test_external_whole_app_scan` runs `~/Dev` end-to-end scans in each cache mode; run its explicit functions only when that corpus is available.
+- To investigate one `~/Dev` extraction reference, run a single `test_external_video_extraction_regression` data row, for example `test_check_refvidparams_nocache:20150727_115225.mp4`.
 - Prefer targeted `ctest --test-dir QtProject/builds/build-debug-6.10.1-macos ...` invocations when using CTest on macOS; the test CMake config chooses a platform plugin compatible with the current Qt build.
-- Do not run `test_local_large_video_corpus` unless explicitly requested; its active functions require the mounted 100GB folder.
+- Do not run `test_external_large_video_corpus` unless explicitly requested; its active functions require the mounted 100GB folder.
 - Package macOS binaries: `npm run binaries`
 - Rebuild vendored macOS deps only when needed: `npm run qt-macos`, `npm run ffmpeg-macos`, `npm run opencv-macos`
