+++ QtProject/tests/repo/auto_delete/tst_auto_delete.cpp
@@ -9,12 +9,12 @@
 
 #include <functional>
 
-#include "../../app/comparison/comparison.h"
-#include "../../app/comparison/ui_comparison.h"
-#include "../../app/db.h"
-#include "../../app/mainwindow.h"
-#include "../../app/prefs.h"
-#include "../../app/ui_mainwindow.h"
+#include "../../../app/comparison/comparison.h"
+#include "../../../app/comparison/ui_comparison.h"
+#include "../../../app/db.h"
+#include "../../../app/mainwindow.h"
+#include "../../../app/prefs.h"
+#include "../../../app/ui_mainwindow.h"
 
 class TestAutoDelete : public QObject
 {
